Fix type resolution for redundant qualified paths
More tests adjustments
